picture-recognition/requirements.txt
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
numpy==1.24.3
matplotlib==3.7.1
scikit-learn==1.2.2
tqdm==4.65.0
pandas==2.0.1